vmm: Move Vcpu::configure() to arch crate
