+++ arch/src/aarch64/mod.rs
@@ -5,7 +5,16 @@ pub mod layout;
 
 use crate::RegionType;
 use kvm_ioctls::*;
-use vm_memory::{GuestAddress, GuestMemoryMmap, GuestUsize};
+use vm_memory::{GuestAddress, GuestMemoryAtomic, GuestMemoryMmap, GuestUsize};
+
+#[derive(Debug)]
+pub enum Error {}
+
+impl From<Error> for super::Error {
+    fn from(e: Error) -> super::Error {
+        super::Error::AArch64Setup(e)
+    }
+}
 
 /// Stub function that needs to be implemented when aarch64 functionality is added.
 pub fn arch_memory_regions(size: GuestUsize) -> Vec<(GuestAddress, usize, RegionType)> {
@@ -18,6 +27,15 @@ pub fn arch_memory_regions(size: GuestUsize) -> Vec<(GuestAddress, usize, Region
 pub struct EntryPoint {
     /// Address in guest memory where the guest must start execution
     pub entry_addr: GuestAddress,
+}
+
+pub fn configure_vcpu(
+    _fd: &VcpuFd,
+    _id: u8,
+    _kernel_entry_point: Option<EntryPoint>,
+    _vm_memory: &GuestMemoryAtomic<GuestMemoryMmap>,
+) -> super::Result<()> {
+    unimplemented!();
 }
 
 /// Stub function that needs to be implemented when aarch64 functionality is added.
